+++ src/ResourceManager/AzureBatch/ChangeLog.md
@@ -19,6 +19,10 @@
 -->
 ## Current Release
 
+## Version 4.1.0
+* Release new cmdlet Get-AzureBatchPoolNodeCounts
+* Release new cmdlet Start-AzureBatchComputeNodeServiceLogUpload
+
 ## Version 4.0.7
 * Set minimum dependency of module to PowerShell 5.0
 * Updated New-AzureBatchPool documentation to remove deprecated example